+++ src/CarListingApp.Services/DTOs/Car/CarDto.cs
@@ -6,7 +6,7 @@ public class CarDto
 {
     public int Id { get; set; }
     
-    public double Price { get; set; }
+    public decimal Price { get; set; }
 
     public string Brand { get; set; } = null!;
 
@@ -18,9 +18,9 @@ public class CarDto
 
     public string? Vin { get; set; }
 
-    public double? EngineDisplacement { get; set; }
+    public decimal? EngineDisplacement { get; set; }
 
-    public double? EnginePower { get; set; }
+    public int? EnginePower { get; set; }
 
     public int? Mileage { get; set; }
